+++ caffe2/utils/fatal_signal_asan_no_sig_test.cc
@@ -1,5 +1,5 @@
 #include "caffe2/utils/signal_handler.h"
-#if defined(CAFFE2_SUPPORTS_FATAL_SIGNAL_HANDLERS)
+#if defined(C10_SUPPORTS_FATAL_SIGNAL_HANDLERS)
 #include <gtest/gtest.h>
 #include <pthread.h>
 #include <unistd.h>
@@ -145,4 +145,4 @@ TEST(fatalSignalTest, SIGSEGV8) {
 TEST(fatalSignalTest, SIGABRT8_NOPRINT) {
   TEST_FATAL_SIGNAL_NO_PRINT(SIGABRT, "SIGABRT", 8);
 }
-#endif // defined(CAFFE2_SUPPORTS_FATAL_SIGNAL_HANDLERS)
+#endif // defined(C10_SUPPORTS_FATAL_SIGNAL_HANDLERS)
