+++ caffe2/utils/signal_handler.h
@@ -1,45 +1,24 @@
 #pragma once
 
-#include "caffe2/core/common.h"
-
-#if defined(__APPLE__)
-#define CAFFE2_SUPPORTS_SIGNAL_HANDLER
-#elif defined(__linux__) && !defined(CAFFE2_DISABLE_SIGNAL_HANDLERS)
-#define CAFFE2_SUPPORTS_FATAL_SIGNAL_HANDLERS
-#define CAFFE2_SUPPORTS_SIGNAL_HANDLER
-#endif
+#include <c10/util/signal_handler.h>
 
 namespace caffe2 {
 
-class TORCH_API SignalHandler {
+#if defined(C10_SUPPORTS_FATAL_SIGNAL_HANDLERS)
+class TORCH_API C2FatalSignalHandler : public c10::FatalSignalHandler {
  public:
-  enum class Action {
-    NONE,
-    STOP
-  };
-
-  // Constructor. Specify what action to take when a signal is received.
-  SignalHandler(Action SIGINT_action,
-                Action SIGHUP_action);
-  ~SignalHandler();
-
-  Action CheckForSignals();
+  void fatalSignalHandlerPostProcess() override;
+  static C2FatalSignalHandler& getInstance();
 
  private:
-  bool GotSIGINT();
-  bool GotSIGHUP();
-  Action SIGINT_action_;
-  Action SIGHUP_action_;
-  unsigned long my_sigint_count_;
-  unsigned long my_sighup_count_;
+  explicit C2FatalSignalHandler();
 };
 
-#if defined(CAFFE2_SUPPORTS_FATAL_SIGNAL_HANDLERS)
 // This works by setting up certain fatal signal handlers. Previous fatal
 // signal handlers will still be called when the signal is raised. Defaults
 // to being off.
 TORCH_API void setPrintStackTracesOnFatalSignal(bool print);
 TORCH_API bool printStackTracesOnFatalSignal();
-#endif // defined(CAFFE2_SUPPORTS_SIGNAL_HANDLER)
+#endif // defined(C10_SUPPORTS_SIGNAL_HANDLER)
 
-}  // namespace caffe2
+} // namespace caffe2
