+++ torch/csrc/jit/python/init.cpp
@@ -93,6 +93,7 @@
 #include <torch/csrc/jit/tensorexpr/tensorexpr_init.h>
 
 #include <c10/macros/Export.h>
+#include <c10/util/signal_handler.h>
 #include <caffe2/serialize/inline_container.h>
 
 #include <ATen/core/function_schema.h>
@@ -1356,6 +1357,11 @@ void initJITBindings(PyObject* module) {
 
   m.def("_jit_assert_is_instance", [](py::object obj, const TypePtr& type) {
     toIValue(std::move(obj), type);
+  });
+
+  m.def("set_print_stack_traces_on_fatal_signal", [](bool print) {
+    c10::FatalSignalHandler::getInstance().setPrintStackTracesOnFatalSignal(
+        print);
   });
 
   initPythonCustomClassBindings(module);
